version(2.1.3) feat(installMany) (#6)
`installMany` allows you to write the following type of code:

    installMany(router, routes)

where router might be

    {
        "signup": Router.post<SignupRequest, SignupResponse>('/api/auth/signup'),
        "login": Router.post<LoginRequest, LoginResponse>('/api/auth/login'),
    }

and routes might be

    {
        "signup": ... as (req: Request, res: ExpressResponse) => Promise<Result<SignupResponse>>,
        "login": ... as (req: Request, res: ExpressResponse) => Promise<Result<LoginResponse>>,
    }

Note that, even though the objects share the same "spine" (their keys are identical), their values have extremely different types. The bulk of this method is the type signature, which carefully assumes nothing about the objects and then pattern matches into the values using `infer`, `keyof`, and `extends` from the recent TypeScript "conditional types" keywords package.

Note also that the obvious thing of having `TRoutes extends Record<string, unknown>` does not work because that makes us unable to write `routes[key].install(...)` at the end of the function.

Note additionally that we have to infer both `TRequest` and `TResponse` in the type of `routes` ause simply saying `[K in keyof TRoutes]: RouteClient<infer T, infer U>` fixes the _same_ T and U for all values of that object.