feat(key errors) (#9)
This bumps the version to 2.1.4. The idea is that frontend/mobile clients shouldn't have to do string comparison on an English sentence to figure out which error they received. The "key" is yet another string, but it's designed to be a symbol (like a Ruby :symbol). For example `key: "unauthorized"` when `error: "Incorrect username or password"`.